pyproject.toml
[tool.poetry]
name = "b2broker-test"
version = "0.1.0"
description = ""
authors = ["Ramil Aglyautdinov <aglyautdinov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
python-decouple = "^3.8"
mysqlclient = "^2.2.4"
djangorestframework = "^3.15.1"
djangorestframework-jsonapi = "^6.1.0"
django-filter = "^24.2"
drf-nested-routers = "^0.93.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
